requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
databases[postgresql]==0.9.0
asyncpg==0.29.0
pytz==2024.1  # или перейти на zoneinfo (см. ниже)
